+++ js/main.js
@@ -1,7 +1,8 @@
 import './form-validate.js';
-import './map.js';
-import './api.js';
+import './main-pin-marker.js';
 import './utils/reset-map.js';
 import './utils/send-form.js';
-import './load-map-data.js';
-
+import './load-map.js';
+import './pin-marker.js';
+import './sort-pin.js';
+import './api.js';
