+++ js/utils/send-form.js
@@ -1,7 +1,7 @@
 import {adForm} from '../form-validate.js';
 import {request} from '../api.js';
-import {showFormSuccessAlert} from './form-success-alert.js';
-import {showFormErrorAlert} from './form-error-alert.js';
+import {showFormSuccessAlert} from './form-success-message.js';
+import {showFormErrorAlert} from './form-error-message.js';
 
 const setFormSubmit = () => {
   const onSuccess = () => {
